tutorials/hello/hello_app/test/unit/modules/hello/commands/test_command.spec.ts
import { TestCommand } from '../../../../../src/app/modules/hello/commands/test_command';

describe('TestCommand', () => {
  let command: TestCommand;

	beforeEach(() => {
		command = new TestCommand();
	});

	describe('constructor', () => {
		it('should have valid name', () => {
			expect(command.name).toEqual('test');
		});

		it('should have valid schema', () => {
			expect(command.schema).toMatchSnapshot();
		});
	});

	describe('verify', () => {
		describe('schema validation', () => {
      it.todo('should throw errors for invalid schema');
      it.todo('should be ok for valid schema');
    });
	});

	describe('execute', () => {
    describe('valid cases', () => {
      it.todo('should update the state store');
    });

    describe('invalid cases', () => {
      it.todo('should throw error');
    });
	});
});
